mus: Move to own pkg
